Move AppContainer and MainScreen to @wordpress/edit-post as Editor and Layout
