--- src/app/style.scss
@@ -1,6 +1,0 @@
-
-.container {
-	flex: 1;
-	justify-content: flex-start;
-	background-color: #fff;
-}
